third_party/WebKit/LayoutTests/fast/dom/shadow/select-in-shadowdom-expected.txt

CONSOLE WARNING: line 18: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
When selecting from non-anchor Node to anchor node in ShadowDOM, client should not cause page jump.

Selecting from a node to another node in ShadowDOM. This should not start page navigation.
PASS lastClickTarget is null


Clicking a node in ShadowDOM.
PASS lastClickTarget is host


PASS successfullyParsed is true

TEST COMPLETE
